requirements.txt
python-dotenv
azure-identity
agent-framework==1.0.0rc3
opentelemetry-semantic-conventions-ai==0.4.13
openai>=1.40.0